% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{build_ctx}
\alias{build_ctx}
\title{Build a validation context object for a given project}
\usage{
build_ctx(project_path, rendered_pdf = NULL)
}
\arguments{
\item{project_path}{Path to the project root (must contain _quarto.yml and
index.qmd).}

\item{rendered_pdf}{Optional path to a rendered PDF file.}
}
\value{
A named list used as the context argument to each rule's check
function.
}
\description{
Build a validation context object for a given project
}
\keyword{internal}
