% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qmd.R
\name{classify_qmd_file}
\alias{classify_qmd_file}
\title{Classify a .qmd/.bib file by role}
\usage{
classify_qmd_file(path)
}
\arguments{
\item{path}{File path or filename.}
}
\value{
One of "prelim", "body", "appendix", "bibliography".
}
\description{
Classify a .qmd/.bib file by role
}
\keyword{internal}
