% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_project}
\alias{locate_project}
\title{Find the nearest ancestor containing _quarto.yml}
\usage{
locate_project(start = ".")
}
\arguments{
\item{start}{Starting directory (defaults to working directory).}
}
\value{
Absolute path to the project root.
}
\description{
Find the nearest ancestor containing _quarto.yml
}
\keyword{internal}
