% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qmd.R
\name{parse_qmd_yaml}
\alias{parse_qmd_yaml}
\title{Parse the YAML front-matter of a .qmd file}
\usage{
parse_qmd_yaml(path)
}
\arguments{
\item{path}{Path to a .qmd file.}
}
\value{
A list, possibly empty.
}
\description{
Parse the YAML front-matter of a .qmd file
}
\keyword{internal}
