% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_report.R
\name{render_check_report}
\alias{render_check_report}
\title{Render a check report}
\usage{
render_check_report(report, format = c("console", "markdown", "json"))
}
\arguments{
\item{report}{A \code{uomthesis_check_report}.}

\item{format}{"console" | "markdown" | "json".}
}
\value{
Invisible character (markdown/json) or report (console).
}
\description{
Render a check report
}
\keyword{internal}
