% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_copyright_text}
\alias{rule_copyright_text}
\title{Rule: copyright.tex must contain all four policy copyright bullets}
\usage{
rule_copyright_text()
}
\value{
A rule spec list.
}
\description{
Rule: copyright.tex must contain all four policy copyright bullets
}
\keyword{internal}
