% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_csl_bundled_or_path_exists}
\alias{rule_csl_bundled_or_path_exists}
\title{Rule: csl setting must be a bundled style or an existing file path}
\usage{
rule_csl_bundled_or_path_exists()
}
\value{
A rule spec list.
}
\description{
Rule: csl setting must be a bundled style or an existing file path
}
\keyword{internal}
