% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_declaration_text}
\alias{rule_declaration_text}
\title{Rule: declaration.tex must contain the policy declaration text}
\usage{
rule_declaration_text()
}
\value{
A rule spec list.
}
\description{
Rule: declaration.tex must contain the policy declaration text
}
\keyword{internal}
