% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_english_language}
\alias{rule_english_language}
\title{Rule: lang must be English when set}
\usage{
rule_english_language()
}
\value{
A rule spec list.
}
\description{
Rule: lang must be English when set
}
\keyword{internal}
