% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_font_allowed}
\alias{rule_font_allowed}
\title{Rule: mainfont must be in the allowed set}
\usage{
rule_font_allowed()
}
\value{
A rule spec list.
}
\description{
Rule: mainfont must be in the allowed set
}
\keyword{internal}
