% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_font_engine_compat}
\alias{rule_font_engine_compat}
\title{Rule: mainfont must be pdflatex-safe when pdf-engine is pdflatex}
\usage{
rule_font_engine_compat()
}
\value{
A rule spec list.
}
\description{
Rule: mainfont must be pdflatex-safe when pdf-engine is pdflatex
}
\keyword{internal}
