% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_linespacing_allowed}
\alias{rule_linespacing_allowed}
\title{Rule: linestretch must be 1.5 or 2.0 when set}
\usage{
rule_linespacing_allowed()
}
\value{
A rule spec list.
}
\description{
Rule: linestretch must be 1.5 or 2.0 when set
}
\keyword{internal}
