% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_prelim_order}
\alias{rule_prelim_order}
\title{Rule: prelim chapters must appear before body chapters before appendices}
\usage{
rule_prelim_order()
}
\value{
A rule spec list.
}
\description{
Rule: prelim chapters must appear before body chapters before appendices
}
\keyword{internal}
