% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{rule_thesis_format}
\alias{rule_thesis_format}
\title{Rule: thesis_format must be "standard" or "journal"}
\usage{
rule_thesis_format()
}
\value{
A rule spec list.
}
\description{
Rule: thesis_format must be "standard" or "journal"
}
\keyword{internal}
