% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uomthesis-package.R
\docType{package}
\name{uomthesis-package}
\alias{uomthesis}
\alias{uomthesis-package}
\title{uomthesis: University of Manchester AMBS PhD Thesis Template and Compliance Tools}
\description{
Provides a Quarto template (standard and journal formats), a project scaffolder, a word counter, and a compliance validator for PhD theses submitted to Alliance Manchester Business School (AMBS) in the Faculty of Humanities at the University of Manchester. Encodes the requirements of the University's Presentation of Theses Policy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabbocg/uomthesis}
  \item \url{https://gabbocg.github.io/uomthesis}
  \item Report bugs at \url{https://github.com/gabbocg/uomthesis/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Cabrera \email{gabriel.cabrera.guz@gmail.com}

}
\keyword{internal}
